+++ sdk/ai/Azure.AI.Projects.Agents/tests/AgentsTests.cs
@@ -1077,6 +1077,47 @@ public class AgentsTests : AgentsTestBase
         Assert.That(backwards[1].Id, Is.EqualTo(records[records.Count - 3].Id));
     }
 
+    [RecordedTest]
+    public async Task TestDraftAgentVersion()
+    {
+        AgentAdministrationClient agentsClient = GetTestClient();
+        DeclarativeAgentDefinition agentDefinition = new(TestEnvironment.FOUNDRY_MODEL_NAME);
+        ProjectsAgentVersion newAgentVersion = await agentsClient.CreateAgentVersionAsync(
+            AGENT_NAME,
+            new ProjectsAgentVersionCreationOptions(agentDefinition)
+            {
+                Metadata = { ["delete_me"] = "please " },
+            }
+        );
+        agentDefinition = new(model: TestEnvironment.FOUNDRY_MODEL_NAME)
+        {
+            Instructions = "You are a prompt agent which gives wrong answers with 0.1 probability."
+        };
+        ProjectsAgentVersion agentVersionDraft = await agentsClient.CreateAgentVersionAsync(
+            agentName: AGENT_NAME,
+            options: new(agentDefinition)
+            {
+                Draft = true
+            }
+        );
+        Assert.That(agentVersionDraft.Name, Is.EqualTo(AGENT_NAME));
+        Assert.That(agentVersionDraft.Version.StartsWith("draft"), Is.True);
+        Assert.That(agentVersionDraft.Definition, Is.InstanceOf<DeclarativeAgentDefinition>());
+        Assert.That(((DeclarativeAgentDefinition)agentVersionDraft.Definition).Instructions, Is.EqualTo("You are a prompt agent which gives wrong answers with 0.1 probability."));
+        // Blocked by work ADO item 5447664
+        //ProjectsAgentRecord record = await agentsClient.GetAgentAsync(AGENT_NAME);
+        //Assert.That(record.GetLatestVersion().Version, Is.EqualTo(newAgentVersion.Version));
+        HashSet<string> versions = [.. await agentsClient.GetAgentVersionsAsync(agentName: AGENT_NAME).Select(x => x.Version).ToListAsync()];
+        Assert.That(versions, Does.Contain(newAgentVersion.Version));
+        Assert.That(versions, Does.Not.Contain(agentVersionDraft.Version));
+        versions = [.. await agentsClient.GetAgentVersionsAsync(agentName: AGENT_NAME, includeDrafts: false).Select(x => x.Version).ToListAsync()];
+        Assert.That(versions, Does.Contain(newAgentVersion.Version));
+        Assert.That(versions, Does.Not.Contain(agentVersionDraft.Version));
+        versions = [.. await agentsClient.GetAgentVersionsAsync(agentName: AGENT_NAME, includeDrafts: true).Select(x => x.Version).ToListAsync()];
+        Assert.That(versions, Does.Contain(newAgentVersion.Version));
+        Assert.That(versions, Does.Contain(agentVersionDraft.Version));
+    }
+
     #region Helpers
     public static async Task DeleteAllSessionsAsync(AgentAdministrationClient agentsClient, string agentName)
     {